Addition of 64-bit Rasberry Pi OS support
In the current version of the driver, the pointer length of "Source_Buffer_Addr" is hardcoded as UDOUBLE (uint32_t),
meaning the use of a non-32-bit OS will result in a Segmentation fault.
This is why I have changed the type to UBYTE* (uint8_t*).
As this address is used only internally, the existing code will still work.